+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Changed: Find Path to not return a path that contains one or more vertices that is can't be retrieved because of visibility restrictions
 * Changed: Reduced DefaultIndexSelectionStrategy cache load time from 1hr to 5min
 * Added: Added a hasId method to the Query class to allow searches to be filtered by element ID.
+* Fix: Extended data element type value for edges
 
 # v3.0.0
 * Changed: Removed ES 2 support and replaced it with ES 5 support